+++ RayTracingWeekend/noise.cpp
@@ -1,6 +1,6 @@
 #include "noise.h"
 
-float perlin::ranfloat[perlin::SIZE];
+double perlin::ranfloat[perlin::SIZE];
 vec3 perlin::ranvec[perlin::SIZE];
 int perlin::perm_x[perlin::SIZE];
 int perlin::perm_y[perlin::SIZE];
