+++ RayTracingWeekend/ray.h
@@ -6,17 +6,17 @@ class ray
 {
 public:
 	ray() {};
-	ray(const vec3& a, const vec3 &b, float ti) { _origin = a; _direction = b; _time = ti; }
+	ray(const vec3& a, const vec3 &b, double ti) { _origin = a; _direction = b; _time = ti; }
 
 	vec3 origin() const { return _origin; }
 	vec3 direction() const { return _direction; }
-	vec3 point_at_parameter(float t) const { return _origin + _direction * t; }
+	vec3 point_at_parameter(double t) const { return _origin + _direction * t; }
 
-	float time() const { return _time; }
+	double time() const { return _time; }
 
 private:
 	vec3 _origin;
 	vec3 _direction;
 
-	float _time;
+	double _time;
 };
